feat(suite): add CoinjoinAccountDiscovery and rename folder
