+++ packages/suite/src/components/wallet/WalletLayout/components/CoinjoinAccountDiscovery/AccountLoadingProgress.tsx
