+++ packages/suite/src/components/wallet/WalletLayout/components/CoinjoinAccountDiscovery/CoinjoinAccountDiscoveryProgress.tsx
